integrate new nflog
