+++ enforcer/proxy/enforcerproxy_test.go
@@ -89,14 +89,14 @@ func createPUInfo() *policy.PUInfo {
 			Address:  "192.30.253.0/24",
 			Port:     "80",
 			Protocol: "TCP",
-			Action:   policy.Reject,
+			Policy:   &policy.FlowPolicy{Action: policy.Reject},
 		},
 
 		policy.IPRule{
 			Address:  "192.30.253.0/24",
 			Port:     "443",
 			Protocol: "TCP",
-			Action:   policy.Accept,
+			Policy:   &policy.FlowPolicy{Action: policy.Accept},
 		},
 	}
